Hold the random number generator in one place.
Instead of each class that needs random number generators having
its own reference, just have the model class, and pass it around.
This means the system is a bit closer to being stateless, which
seems good, and can see the random state from the model object.
It feels like it should belong to it.